requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
anyio==4.7.0
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
asyncio==3.4.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
djando==0.2
Django==5.1.4
djangorestframework==3.15.2
executing==2.1.0
fonttools==4.54.1
frozenlist==1.5.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
influxdb-client==1.48.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.0
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
nest-asyncio==1.6.0
numpy==2.1.2
optree==0.13.0
packaging==24.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.0.0
psycopg==3.2.3
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.9
pytz==2024.2
pywin32==307
pyzmq==26.2.0
reactivex==4.0.4
requests==2.32.3
rich==13.9.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
stack-data==0.6.3
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.18.3
